backend/requirements.txt
beautifulsoup4
playwright
requests
psycopg
python-dotenv
psycopg2-binary
flask
flask_sqlalchemy
